seishub event converter: skipping download of existing files

+++ seishub_event_format_to_quakeml/fetch_all_seishub_events.sh
@@ -1,6 +1,10 @@
 #!/bin/sh
-mkdir all_resources
+mkdir -p all_resources
 for FILE in `cat all_resource_names`
 do
+    if [ ! -f "all_resources/$FILE" ]
+    then
+        break
+    fi
     wget -O all_resources/$FILE http://teide.geophysik.uni-muenchen.de:8080/xml/seismology/event/$FILE
 done
